split api tests
